requirements.txt
flask==3.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1 